Revert "Prevent starvation"
This reverts commit 126072acaf74159637d9610084899f0118f4e94b.

+++ src/bq_common/platform/thread/spin_lock.h
@@ -220,7 +220,7 @@ namespace bq {
 
         public:
             spin_lock_rw_crazy()
-                : state_(static_cast<uint64_t>(0))
+                : state_(0)
             {
             }
